Fix 15 audit bugs + UX improvements
Seeker: fix applied stat capped at 10 (use separate count query), fix
formatSalary for sub-₹1000 values in saved jobs, verify seeker actually
applied before showing /applied success page.

Employer: fix 'Browse Candidates' link pointing to /jobs instead of
/employer/candidates, fix verification 'View' link to include /applicants,
add WALK_IN job type to wizard, fix experienceMin '6m' stored as 6 years
(use parseFloat + value '0.5'), add salary min/max validation in wizard.

UX: all 'Post a Job' links now use ?role=employer to skip role selection
screen, 'Job seeker login' link uses ?role=seeker to skip to OTP step,
mobile filter drawer added to jobs page so filters work on small screens,
fix daily wage label in job detail sidebar, fix mobile My Jobs anchor.

+++ src/app/api/employer/post-job/route.ts
@@ -60,7 +60,7 @@ export async function POST(req: NextRequest) {
       salaryMax: salaryMax ? parseInt(salaryMax) : null,
       salaryUnit: salaryUnit ?? "monthly",
       vacancies: parseInt(vacancies) || 1,
-      experienceMin: parseInt(experienceMin) || 0,
+      experienceMin: parseFloat(experienceMin) || 0,
       pincode: pincode?.trim() || null,
       qualificationRequired: qualificationRequired ?? null,
       status: "PENDING_REVIEW",
